+++ src/architect/account/account.utils.ts
@@ -2,19 +2,18 @@ import { Flags } from '@oclif/core';
 import chalk from 'chalk';
 import inquirer from 'inquirer';
 import AppService from '../../app-config/service';
+import * as Docker from '../../common/utils/docker';
 import Account from './account.entity';
 
 export default class AccountUtils {
   static flags = {
-    account: {
-      non_sensitive: true,
-      ...Flags.string({
-        description: 'Architect account',
-        env: 'ARCHITECT_ACCOUNT',
-        char: 'a',
-        parse: async value => value.toLowerCase(),
-      }),
-    },
+    account: Flags.string({
+      description: 'Architect account',
+      env: 'ARCHITECT_ACCOUNT',
+      char: 'a',
+      parse: async value => value.toLowerCase(),
+      sensitive: false,
+    }),
   };
 
   static getLocalAccount(): Account {
@@ -23,6 +22,18 @@ export default class AccountUtils {
       id: 'dev',
       name: 'dev (Local Machine)',
     };
+  }
+
+  static async shouldListLocalAccounts(exit_on_failure?: boolean): Promise<boolean> {
+    try {
+      await Docker.verify();
+      return true;
+    } catch (e: any) {
+      if (exit_on_failure) {
+        process.exit();
+      }
+      return false;
+    }
   }
 
   static isLocalAccount(account: Account): boolean {
@@ -51,7 +62,7 @@ export default class AccountUtils {
     const token_json = await app.auth.getPersistedTokenJSON();
     if (!token_json || token_json.email === 'unknown') {
       console.log(chalk.yellow('In order to access remote accounts you can login by running `architect login`'));
-      let account: Account;
+      await this.shouldListLocalAccounts(true);
       const answers: { account: Account } = await inquirer.prompt([
         {
           type: 'list',
@@ -72,6 +83,7 @@ export default class AccountUtils {
       }
     } else {
       inquirer.registerPrompt('autocomplete', require('inquirer-autocomplete-prompt'));
+      const can_run_local = await this.shouldListLocalAccounts();
       const answers: { account: Account } = await inquirer.prompt([
         {
           type: 'autocomplete',
@@ -81,7 +93,7 @@ export default class AccountUtils {
           source: async (answers_so_far: any, input: string) => {
             const { data } = await app.api.get('/accounts', { params: { q: input, limit: 10 } });
             const accounts = data.rows as Account[];
-            if (options?.ask_local_account) {
+            if (options?.ask_local_account && can_run_local) {
               accounts.unshift(this.getLocalAccount());
             }
             return accounts.map((a) => ({ name: a.name, value: a }));