+++ src/commands/environments/create.ts
@@ -16,27 +16,26 @@ interface CreateEnvironmentDto {
 export default class EnvironmentCreate extends BaseCommand {
   static aliases = ['environment:create', 'envs:create', 'env:create'];
   static description = 'Register a new environment with Architect Cloud';
-
+  static examples = [
+    'environment:create --account=myaccount myenvironment',
+    'environment:create --account=myaccount --ttl=5days --description="My new temporary Architect environment" myenvironment',
+  ];
   static flags = {
     ...BaseCommand.flags,
     ...AccountUtils.flags,
     ...PlatformUtils.flags,
-    description: {
-      non_sensitive: true,
-      ...Flags.string({
-        description: 'Environment Description',
-      }),
-    },
-    ttl: {
-      non_sensitive: true,
-      ...Flags.string({
-        description: 'The TTL of the environment in a duration of time, ex. 30d, 12h, or 30m',
-      }),
-    },
+    description: Flags.string({
+      description: 'Environment Description',
+      sensitive: false,
+    }),
+    ttl: Flags.string({
+      description: 'The TTL of the environment in a duration of time, ex. 30d, 12h, or 30m',
+      sensitive: false,
+    }),
   };
 
   static args = [{
-    non_sensitive: true,
+    sensitive: false,
     name: 'environment',
     description: 'Name to give the environment',
     parse: async (value: string): Promise<string> => value.toLowerCase(),