+++ src/commands/logs.ts
@@ -17,51 +17,44 @@ export default class Logs extends BaseCommand {
   }
 
   static description = 'Get logs from services both locally and remote';
-
+  static examples = [
+    'architect logs',
+    'architect logs --follow --raw --timestamps',
+  ];
   static flags = {
     ...BaseCommand.flags,
     ...AccountUtils.flags,
     ...EnvironmentUtils.flags,
-    follow: {
-      non_sensitive: true,
-      ...Flags.boolean({
-        description: 'Specify if the logs should be streamed.',
-        char: 'f',
-        default: false,
-      }),
-    },
-    since: {
-      non_sensitive: true,
-      ...Flags.string({
-        description: 'Only return logs newer than a relative duration like 5s, 2m, or 3h. Defaults to all logs. Only one of since-time / since may be used.',
-        default: '',
-      }),
-    },
-    raw: {
-      non_sensitive: true,
-      ...Flags.boolean({
-        description: 'Show the raw output of the logs.',
-        default: false,
-      }),
-    },
-    tail: {
-      non_sensitive: true,
-      ...Flags.integer({
-        description: 'Lines of recent log file to display. Defaults to -1 with no selector, showing all log lines otherwise 10, if a selector is provided.',
-        default: -1,
-      }),
-    },
-    timestamps: {
-      non_sensitive: true,
-      ...Flags.boolean({
-        description: 'Include timestamps on each line in the log output.',
-        default: false,
-      }),
-    },
+    follow: Flags.boolean({
+      description: 'Specify if the logs should be streamed.',
+      char: 'f',
+      default: false,
+      sensitive: false,
+    }),
+    since: Flags.string({
+      description: 'Only return logs newer than a relative duration like 5s, 2m, or 3h. Defaults to all logs. Only one of since-time / since may be used.',
+      default: '',
+      sensitive: false,
+    }),
+    raw: Flags.boolean({
+      description: 'Show the raw output of the logs.',
+      default: false,
+      sensitive: false,
+    }),
+    tail: Flags.integer({
+      description: 'Lines of recent log file to display. Defaults to -1 with no selector, showing all log lines otherwise 10, if a selector is provided.',
+      default: -1,
+      sensitive: false,
+    }),
+    timestamps: Flags.boolean({
+      description: 'Include timestamps on each line in the log output.',
+      default: false,
+      sensitive: false,
+    }),
   };
 
   static args = [{
-    non_sensitive: true,
+    sensitive: false,
     name: 'resource',
     description: 'Name of resource',
     required: false,